feat: add group pairing
