+++ src/app/(content)/feed/page.tsx
@@ -3,29 +3,11 @@ import Details from "~/components/Details";
 import groupDetails from "~/types";
 import React, { useEffect, useState } from "react";
 import { db } from "~/lib/api/firebaseConfig";
-import { collection, query, onSnapshot, Timestamp } from "firebase/firestore";
+import { collection, query, onSnapshot, Timestamp, setDoc, doc } from "firebase/firestore";
 import { useUser } from "@clerk/nextjs";
 // import { redirect } from "next/dist/server/api-utils";
 import { redirect } from "next/navigation";
-
-function formatDateTime(timestamp: Timestamp): [String | null, String | null] {
-  if (!timestamp) {
-    return [null,null];
-  }
-  const date = timestamp.toDate()
-  const month = String(date.getMonth() + 1).padStart(2, '0');  // getMonth() is 0-indexed (January is 0)
-  const day = String(date.getDate()).padStart(2, '0');
-  const year = String(date.getFullYear()).slice(-2);
-  let hours = date.getHours();
-  const minutes = String(date.getMinutes()).padStart(2, '0');
-
-
-  
-  const ampm = hours >= 12 ? 'PM' : 'AM';
-  hours = hours % 12;
-  hours = hours ? hours : 12; // '0' hour should be '12'
-  return [`${month}/${day}/${year}`, `${hours}:${minutes} ${ampm}`];
-}
+import formatDateTime from "~/helpers/date_helper";
 
 function InClass() {
   const { user } = useUser();
